Move all plugins into /plugins
