+++ code/lib/builder-vite/src/plugins/inject-export-order-plugin.ts
