+++ code/lib/builder-vite/src/plugins/source-loader-plugin.ts
@@ -1,7 +1,7 @@
 import type { Plugin } from 'vite';
 import sourceLoaderTransform from '@storybook/source-loader';
 import MagicString from 'magic-string';
-import type { ExtendedOptions } from './types';
+import type { ExtendedOptions } from '../types';
 
 const storyPattern = /\.stories\.[jt]sx?$/;
 const storySourcePattern = /var __STORY__ = "(.*)"/;
