+++ code/lib/builder-vite/src/vite-config.ts
@@ -9,11 +9,13 @@ import type {
 } from 'vite';
 import viteReact from '@vitejs/plugin-react';
 import { isPreservingSymlinks, getFrameworkName } from '@storybook/core-common';
-import { codeGeneratorPlugin } from './code-generator-plugin';
 import { stringifyProcessEnvs } from './envs';
-import { injectExportOrderPlugin } from './inject-export-order-plugin';
-import { mdxPlugin } from './plugins/mdx-plugin';
-import { stripStoryHMRBoundary } from './plugins/strip-story-hmr-boundaries';
+import {
+  codeGeneratorPlugin,
+  injectExportOrderPlugin,
+  mdxPlugin,
+  stripStoryHMRBoundary,
+} from './plugins';
 import type { ExtendedOptions, EnvsRaw } from './types';
 
 export type PluginConfigType = 'build' | 'development';
